Split out serializable.py file into smaller files
