+++ src/syft/__init__.py
@@ -50,8 +50,8 @@ import syft as sy
 from syft.core import common  # noqa: F401
 
 # Convenience Methods
-from syft.core.common.serializable import _deserialize as deserialize
-from syft.core.common.serializable import _serialize as serialize  # noqa: F401
+from syft.core.common.serde.deserialize import _deserialize as deserialize
+from syft.core.common.serde.serialize import _serialize as serialize  # noqa: F401
 from syft.core.node.common.service.repr_service import ReprMessage  # noqa: F401
 from syft.core.node.device.device import Device  # noqa: F401
 from syft.core.node.device.device import DeviceClient  # noqa: F401
