+++ src/syft/core/common/object.py
@@ -3,8 +3,8 @@ from ...proto.core.common.common_object_pb2 import ObjectWithID as ObjectWithID_
 
 # syft imports (sorted by length)
 from ...decorators.syft_decorator_impl import syft_decorator
-from .serializable import _deserialize
-from .serializable import Serializable
+from syft.core.common.serde.deserialize import _deserialize
+from syft.core.common.serde.serializable import Serializable
 from .uid import UID
