+++ src/syft/core/store/storeable_object.py
@@ -2,8 +2,8 @@ from dataclasses import dataclass
 from typing import List, Optional
 
 from ...proto.core.store.store_object_pb2 import StorableObject as StorableObject_PB
-from ..common.serializable import Serializable
-from ..common.serializable import _deserialize
+from syft.core.common.serde.serializable import Serializable
+from syft.core.common.serde.deserialize import _deserialize
 from ..common.uid import UID
